A lot
Added default tooltips through HTML "title" attribute to some not-so-obvious UI elements. Fixed the menu hamburger/close buttons so that it only needs one button (and it looks better too). Added saving/loading of settings, and auto-loading of a Steam installation (that was hell). Also updated w3.css to the latest version as of Feb 25th, 2017 (old version is in app/resources/w3.css.old).