Way more
Switched to using HTML5 color picker for background and text colors. Fixed a bunch of small bugs: addForcedDownload was still failing.
Tweaked the menu - still trying to fix it's issue with the cursor while hovering; it's now slightly smaller, as is it's text.
Also officially fixed the list sizing.
Added a bunch of checking for possible error conditions.
The "settings" value was being saved wrong, causing it to be written to config as both the settings object and a set of objects for the settings values.
Fixed the tabbing while the Launch Options modal is open so that it won't let the user tab back to the program.