License, upgrades, refinement, bugfixes
Added MIT license.

Added "tabs" to about page, split into "built with" and "powered by". Also changed Badger.CSS to use badger-container as a class rather than ID.

Refined: Function to toggle menu on/off. Filtering function - now it won't try to filter ?, +, *, etc other characters that were being recognized as parts of a regexp and breaking the filtering.

Fixed some bugs - launchOptions not saving properly, addForcedDownload would try to work with no steamPath set, addForcedDownload was not immediately adding the added download to the "Steam Apps" list, loadSettings was not properly upgrading settings to v1.2, and the filtering was not updating the number of items in the "Steam Apps" or "Blacklist" lists.