Exit on launch, color fixing
There is now an option to exit RV on launch of Steam.

When opening launch options with a light text color it's hard to see; now it will toggle the text color of the launch options modal to black on open if it's set to a light color, and back to normal once it closes.