Folders can't be short (yet), skin change fix
Changing the displayed value of the folder changes the saved value..need to workaround that to be able to display short folder names.

Fixed issue with setting skin to default on Windows - val (registry value) is set to null for SkinV4 when it's value is empty, and so there is no val.value to test against.