Fixes, updates
Fixed: Select All/Cut/Copy/Paste now work on Mac again (had to add them back in via the menu). Potentially fixed the issue where clicking a label[for="id"] and having it activate the element it's for was not working on Windows. Made some space between the Steam App Controls Apply & Add To Blacklist buttons and the controls above them.

Updates: A lot more work on the README. Added "repository" and "license" fields to package.json files.