Bugfixes, updates; working towards v1.2.2
Trying to fix the Linux .Desktop file categories entry issue. Updated Electron, electron-builder, and w3.css to their latest versions (1.75, 19.22.1, 4.0.4) from NPM. Fixed some bugs including the menu (updating w3.css broke it), a crash when trying to set a random skin but nothing exists. Also updated all the buttons so they now have a hover color and are all w3-button rather than a mixture of that and w3-btn. Also added a bottom-border to the menu items when focused so that it's easier to tell. Finally, fixed the README a little - it was using some incorrect badges.