Some updates, some fixes
Updated icons to include a 1024x1024 png for Linux. Fixed an issue on Linux from Valve apparently modifying the Steam install folder configuration in the last few months. Fixed an issue where the "Launch" functionality would launch Steam using the currently selected options, not those that are saved to the settings. Fixed "Launch Options" -- can't use the "-command" versions any more, apparently. Changed launch options to be radio buttons instead of checkboxes because only one of the currently-available options can be used at a time; also added a "nothing" option which isn't even checked for and so will trigger the default: "//open/main" (open to the user's main Steam window). Updated the "description" to not only say RV is useful in offline mode because it can do things for the Steam client without it needing to be launched in online mode. Implemented a function to check if the program can access a path; also using that to check if the program can access the Steam folder path with RWX permissions. Officially moved to Yarn -- thank you, develar!