Bugfixes, fixes, optimization
Fixed the Launch Options "Reset" button actually resetting them to what is saved rather than the default (everything off). Also set "Nothing" as the default option to be chosen. Removed "Developer" launch option because it's apparently no longer supported... Added another pixel of border on the right of the menu toggle button so it matches the bottom. Made contributor name a link with the title being the link so it can be seen by a user before being clicked.