Starting to make it work, again...
Fixed some of the loading code; it can now load the common Steam data (after the path is set and Load is pressed) and part of the user data (when a user is set), but is failing on localconfig.vdf. Most of the data should also be properly updated in the UI now.